feat: reusable component for applications list
